Move date utils to dates package
