+++ packages/dates/src/utils/addSub.ts
@@ -12,8 +12,6 @@ import {
 	toDate,
 } from 'date-fns';
 
-import { toInteger } from '../converters/number';
-
 export type IntervalType =
 	| 'days'
 	| 'hours'
@@ -46,7 +44,8 @@ export const add = (interval: IntervalType, date: Date | number, amount: number)
 
 export const sub = (interval: IntervalType, dirtyDate: Date | number, dirtyAmount: number): Date => {
 	const func = addMapping[interval];
-	const amount = toInteger(dirtyAmount);
+	let amount = Number(dirtyAmount);
+	amount = amount < 0 ? Math.ceil(amount) : Math.floor(amount);
 	const date = toDate(dirtyDate);
 
 	return func(date, -amount);
