+++ packages/dates/src/utils/index.ts
@@ -1,3 +1,8 @@
+export * from './addSub';
+export { default as areEqual } from './areEqual';
+export { default as diff } from './diff';
+export { default as sort } from './sort';
+export { default as switchTenseForDate } from './switchTenseForDate';
 export { default as isOnOrAfterToday } from './isOnOrAfterToday';
 export { default as isOnOrBeforeToday } from './isOnOrBeforeToday';
 export { default as isOnOrAfterDate } from './isOnOrAfterDate';
