+++ packages/dates/src/utils/misc.ts
@@ -1,8 +1,51 @@
 import { pipe } from 'ramda';
-import { toDate } from 'date-fns';
+import { __ } from '@eventespresso/i18n';
+import { parseISO, toDate } from 'date-fns';
 import { setHours, setMinutes, setSeconds, setYear, setMonth, setDate } from 'date-fns/fp';
 
+import type { OptionsType } from '@eventespresso/adapters';
+
+import { add, sub } from './addSub';
+import type { Intervals, ShiftDateArgs } from './types';
 import type { PrepDatesComparisonFunc } from './types';
+
+export const DATE_INTERVALS: Intervals = {
+	months: __('month(s)'),
+	weeks: __('week(s)'),
+	days: __('day(s)'),
+	hours: __('hour(s)'),
+	minutes: __('minute(s)'),
+};
+
+export const intervalsToOptions = (intervals: Intervals, prependEmpty?: boolean): OptionsType => {
+	const options = Object.entries(intervals).map(([value, label]) => ({ value, label }));
+
+	if (prependEmpty) {
+		return [{ label: '', value: '' }, ...options];
+	}
+	return options;
+};
+
+/**
+ * Shifts the given date according to args.
+ */
+export const shiftDate = (args: ShiftDateArgs) => (date: Date | string): Date => {
+	const parsedDate = date instanceof Date ? date : parseISO(date);
+	if (args?.unit && args?.value && args?.type) {
+		const fn = args.type === 'earlier' ? sub : add;
+		return fn(args.unit, parsedDate, args.value);
+	}
+	return parsedDate;
+};
+
+/**
+ * Sets the default time for a date based on `type`
+ * Default time: 'start' => 8 am, 'end' => 5 pm
+ */
+export const setDefaultTime = (date: Date, type: 'start' | 'end' = 'start'): Date => {
+	const hours = type === 'start' ? 8 : 17;
+	return pipe(setHours(hours), setMinutes(0), setSeconds(0))(date);
+};
 
 /**
  * Sets the time of the date object to zero hour
