+++ packages/dates/src/utils/types.ts
@@ -1,3 +1,15 @@
+import { IntervalType } from './addSub';
+
+export type TzDateFn = (date: Date | string | number, timezone: string) => Date;
+
+export type Intervals = { [key in IntervalType]?: string };
+
+export type ShiftDateArgs = {
+	unit: IntervalType;
+	value: number;
+	type: 'earlier' | 'later';
+};
+
 export type PrepSingleDateComparisonFunc = (firstDate: Date | number, considerTime: boolean) => Date;
 
 export type PrepDatesComparisonFunc = (
