--- packages/utils/src/date/index.ts
@@ -1,7 +1,0 @@
-export * from './addSub';
-export { default as areEqual } from './areEqual';
-export { default as diff } from './diff';
-export { default as sort } from './sort';
-export { default as switchTenseForDate } from './switchTenseForDate';
-export * from './misc';
-export * from './types';
