+++ packages/utils/src/index.ts
@@ -3,7 +3,6 @@ export { createInterpolateElement } from '@wordpress/element';
 export * from './array';
 export * from './boolean';
 export * from './converters';
-export * from './date';
 export * from './dom';
 export * from './function';
 export * from './list';
